Awk Moment When It Actually Hasn't Finished Loading...
Frustrated by my computers inability to update software so that I could
download github and do this assignment, I adjusted Taylor’s SVG
assignment, which shows a tablet that has just downloaded an ebook, to
reflect a moment when “It Actually Hasn't Finished Loading…” To
demonstrate this, I changed the color scheme of the background to a
more calming blue. I added an indefinite animation so that the screen
would flashing from grey to blue. In addition, I added an animation to
the button so that it would spin, also demonstrating an action
associated with “loading.” I attempted to add a gradient to the button;
however, I was unsuccessful in this endeavor.